fixed code, proper file path, organize, and new icons
